// src/app/syllabus-entry/models/CourseTask.ts
export interface CourseTask {
      courseTaskId: number;
      baseSchoolNameId: number;
      courseNameId: number;
      courseName: string,
      courseDurationId: number;
      bnaSubjectNameId: number;
      taskDetail: string;
      remarks:number;
      menuPosition:number;
      isActive: boolean
}